Update bootstrap to be able to run a subset and run a subset when launching an instance.

+++ util/install/ansible-bootstrap.sh
@@ -33,6 +33,10 @@ if [[ -z "${UPGRADE_OS}" ]]; then
   UPGRADE_OS=false
 fi
 
+if [[ -z "${RUN_ANSIBLE}" ]]; then
+  RUN_ANSIBLE=true
+fi
+
 #
 # Bootstrapping constants
 #
@@ -75,9 +79,9 @@ then
 elif grep -q 'Xenial Xerus' /etc/os-release
 then
     SHORT_DIST="xenial"
-else    
+else
     cat << EOF
-    
+
     This script is only known to work on Ubuntu Precise, Trusty and Xenial,
     exiting.  If you are interested in helping make installation possible
     on other platforms, let us know.
@@ -96,7 +100,7 @@ if [ "${UPGRADE_OS}" = true ]; then
     echo "Upgrading the OS..."
     apt-get upgrade -y
 fi
-    
+
 # Required for add-apt-repository
 apt-get install -y software-properties-common python-software-properties
 
@@ -110,7 +114,7 @@ if [[ "precise" = "${SHORT_DIST}" || "trusty" = "${SHORT_DIST}" ]]; then
     apt-key adv --keyserver "${EDX_PPA_KEY_SERVER}" --recv-keys "${EDX_PPA_KEY_ID}"
     add-apt-repository -y "${EDX_PPA}"
 fi
-    
+
 # Install python 2.7 latest, git and other common requirements
 # NOTE: This will install the latest version of python 2.7 and
 # which may differ from what is pinned in virtualenvironments
@@ -136,35 +140,41 @@ PATH=/usr/local/bin:${PATH}
 pip install setuptools=="${SETUPTOOLS_VERSION}"
 pip install virtualenv=="${VIRTUAL_ENV_VERSION}"
 
-# create a new virtual env
-/usr/local/bin/virtualenv "${VIRTUAL_ENV}"
 
-PATH="${PYTHON_BIN}":${PATH}
+if [[ "true" == "${RUN_ANSIBLE}" ]]; then
+    # create a new virtual env
+    /usr/local/bin/virtualenv "${VIRTUAL_ENV}"
 
-# Install the configuration repository to install 
-# edx_ansible role
-git clone ${CONFIGURATION_REPO} ${CONFIGURATION_DIR}
-cd ${CONFIGURATION_DIR}
-git checkout ${CONFIGURATION_VERSION}
-make requirements
+    PATH="${PYTHON_BIN}":${PATH}
 
-cd "${CONFIGURATION_DIR}"/playbooks/edx-east
-"${PYTHON_BIN}"/ansible-playbook edx_ansible.yml -i '127.0.0.1,' -c local -e "configuration_version=${CONFIGURATION_VERSION}"
+    # Install the configuration repository to install
+    # edx_ansible role
+    git clone ${CONFIGURATION_REPO} ${CONFIGURATION_DIR}
+    cd ${CONFIGURATION_DIR}
+    git checkout ${CONFIGURATION_VERSION}
+    make requirements
 
-# cleanup
-rm -rf "${ANSIBLE_DIR}"
-rm -rf "${CONFIGURATION_DIR}"
-rm -rf "${VIRTUAL_ENV}"
-rm -rf "${HOME}/.ansible"
+    cd "${CONFIGURATION_DIR}"/playbooks/edx-east
+    "${PYTHON_BIN}"/ansible-playbook edx_ansible.yml -i '127.0.0.1,' -c local -e "configuration_version=${CONFIGURATION_VERSION}"
 
-cat << EOF
-******************************************************************************
+    # cleanup
+    rm -rf "${ANSIBLE_DIR}"
+    rm -rf "${CONFIGURATION_DIR}"
+    rm -rf "${VIRTUAL_ENV}"
+    rm -rf "${HOME}/.ansible"
 
-Done bootstrapping, edx_ansible is now installed in /edx/app/edx_ansible.
-Time to run some plays.  Activate the virtual env with 
+    cat << EOF
+    ******************************************************************************
 
-> . /edx/app/edx_ansible/venvs/edx_ansible/bin/activate 
+    Done bootstrapping, edx_ansible is now installed in /edx/app/edx_ansible.
+    Time to run some plays.  Activate the virtual env with
 
-******************************************************************************
+    > . /edx/app/edx_ansible/venvs/edx_ansible/bin/activate
+
+    ******************************************************************************
 EOF
+else
+    mkdir -p /edx/ansible/facts.d
+    echo '{ "ansible_bootstrap_run": true }' > /edx/ansible/facts.d/ansible_bootstrap.json
+fi
